+++ spec/fixtures/android_unit_test_checker/TestsINeedThem2.kt
@@ -10,3 +10,23 @@ class TestsINeedThem2AnotherClass
         println("thisNeedsATest")
     }
 }
+
+enum class ProtocolState {
+    WAITING {
+        override fun signal() = TALKING
+    },
+
+    TALKING {
+        override fun signal() = WAITING
+    };
+
+    abstract fun signal(): ProtocolState
+}
+
+value class Password(private val s: String)
+
+sealed class Error(val message: String) {
+    class NetworkError : Error("Network failure")
+    class DatabaseError : Error("Database cannot be reached")
+    class UnknownError : Error("An unknown error has occurred")
+}